move dump to list convo
