+++ cmd/slackdump/internal/list/assets/list_conversation.md
@@ -1,15 +1,16 @@
-# Command Dump
+# Command List Conversation
 
 ## Description
 
-Dump is the mode that allows to dump the following type of conversations:
+List Conversation is the command that allows to dump the following type of
+conversations:
 
 - public and private channels with threads
 - group messages (MPIM)
 - private messages (DMs)
 - individual threads
 
-It downloads file attachments as well, if the `-files` flag is set.
+If the `-files` flag is set, it downloads file attachments as well, 
 
 This is the original low-level mode that applies almost no transformations to
 the API output mode of the Slackdump, and its behaviour would be familiar to
@@ -25,9 +26,9 @@ or URLs or combine file with conversations and individual conversation links.
 
 ## Converting JSON Dumps to Other Formats
 
-To convert the JSON file generated by `slackdump dump` to other formats (for
-example, text), use `slackdump convert`.  Run `slackdump help convert` to
-learn mode.
+To convert the JSON file generated by `slackdump {{ .LongName }}` to other
+formats (for example, text), use `slackdump convert`.  Run `slackdump help
+convert` to learn mode.
 
 For reference, in the old version this function was controlled by `-r` flag.
 
@@ -38,13 +39,13 @@ For reference, in the old version this function was controlled by `-r` flag.
 This command will also enable file download.
 
 ```shell
-slackdump dump -files C051D4052 DHYNUJ00Y
+slackdump {{ .LongName }} -files C051D4052 DHYNUJ00Y
 ```
 
 ### Dump channels listed in my_channels.txt
 
 ```shell
-slackdump dump @my_channels.txt
+slackdump {{ .LongName }} @my_channels.txt
 ```
 
 ### Dump a single thread
@@ -55,14 +56,14 @@ Threads can be specified as a **URL**, or use a Slackdump-specific
 URL:
 
 ```shell
-slackdump dump \
+slackdump {{ .LongName }} \
   https://ora600.slack.com/archives/C051D4052/p1665917454731419
 ```
 
 Slackdump colon notation:
 
 ```shell
-slackdump dump C051D4052:1665917454.731419
+slackdump {{ .LongName }} C051D4052:1665917454.731419
 ```
 
 ### Combined all of the above
@@ -71,7 +72,7 @@ This example shows how you can combine different types of input. URL of the
 thread is omitted for brevity:
 
 ```shell
-slackdump dump -files C051D4052 \
+slackdump {{ .LongName }} -files C051D4052 \
   DHYNUJ00Y \ 
   @my_channels.txt \
   C051D4052:1665917454.731419
