+++ cmd/slackdump/internal/list/dump_wiz.go
@@ -1,4 +1,4 @@
-package dump
+package list
 
 import (
 	"context"
